lesson_29_1/requirements.txt
psycopg2-binary
pytest


